requirements.txt
accelerate==0.34.2
decord==0.6.0
deepspeed==0.12.3
einops==0.8.0
hydra-core==1.3.2
matplotlib==3.9.2
numpy==1.26.4
opencv-python==4.10.0.84
peft==0.8.2
pycocotools==2.0.8
sentencepiece==0.1.99
tensorboard==2.18.0
transformers==4.33.2


